+++ src/main/java/com/example/demo/repository/StateRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 
-public interface StateRepository extends JpaRepository<Estate, Integer> {
+public interface StateRepository extends JpaRepository<Estate, Long> {
 
     @Modifying
     @Query(value = "UPDATE state  SET is_deleted=false WHERE s.id = :id ", nativeQuery = true)
